Tip Servgins je sada int32 ne uint16

+++ models/recipe.go
@@ -5,7 +5,7 @@ type Recipe struct {
 	Ingredients []Ingredient `json:"ingredients"`
 	Steps       []string     `json:"steps"`
 	Image       string       `json:"image"`
-	Servings    uint16       `json:"servings"`
+	Servings    int32        `json:"servings"`
 }
 
 type Ingredient struct {
